voltiq-backend/requirements.txt
# VoltIQ Backend Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
supabase==2.3.4
redis==5.0.1

# ML/AI
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.4.0
joblib==1.3.2
lightgbm==4.3.0
xgboost==2.0.3
tensorflow==2.15.0

# MILP Optimization
ortools==9.8.3296

# TP-Link Tapo Smart Plugs
PyP100==0.1.2

# Auth
PyJWT==2.8.0

# Utilities
python-dotenv==1.0.0
httpx>=0.24.0,<0.26.0
python-multipart==0.0.6
